refactor(frontend): split api/admin/worldViewImport into focused sub-files
Split the monolithic worldViewImport.ts into three focused modules:
- wvImportTreeOps.ts: hierarchy mutations (flatten, prune, rename, reparent,
  auto-resolve, overlap, smart-simplify, AI suggest, etc.)
- wvImportCoverage.ts: GADM coverage analysis, gap detection, union geometry,
  split-deeper, children geometry, coverage-compare endpoints
- wvImportCvMatch.ts: CV color matching SSE pipeline, vision match, mapshape
  match, water review, cluster review, ICP adjustment callbacks

worldViewImport.ts becomes a barrel re-exporting from these three modules
plus retaining core types and basic match operations (lifecycle, match stats,
accept/reject, geoshape, rematch).
